Add rest API definitions for proper swagger client specs.
Currently we do not have a CLI tool to manage projects and tenants in the masterdata-api.

For the time being, we can use the definitions of this package to generate proper swagger clients.

As soon as we have a CLI tool and client authentication, we can probably remove this package and remove managing masterdata entities from the other APIs that use these definitions.